requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=1.5.0
numpy>=1.21.0
langchain>=0.0.350
langchain-community>=0.0.29
langchain-core>=0.1.33
faiss-cpu>=1.7.4
chromadb>=0.4.0
python-docx>=0.8.11
pypdf>=3.17.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
openai>=1.3.0
pillow>=10.0.0
scikit-learn>=1.3.0
filelock>=3.12.0
unstructured>=0.10.30
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tabulate>=0.9.0
openpyxl>=3.1.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=1.10.12
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.30.2
sentence-transformers>=2.2.2
nltk>=3.8.0
tqdm>=4.64.0
evaluate>=0.4.0
rouge-score>=0.1.2
Flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
urllib3>=1.26.0
speechrecognition>=3.10.0
pyaudio>=0.2.11
pyttsx3>=2.90
gtts>=2.3.2
playsound>=1.3.0
pydub>=0.25.1
librosa>=0.10.1
sounddevice>=0.4.6
bert-score>=0.3.13